[Deodhai]: bugs fixed in frame composition timing [Kernel]: now calibrate tsc timing, implemented gettimeofday system call, which returns time in unix format, bug fixed in thread structure where sleep quanta was in 2 byte format now it is in 8 byte format

+++ BaseHdr/Hal/x86_64_sched.h
@@ -126,7 +126,7 @@ typedef struct _au_thread_ {
 	char name[16];
 	uint8_t state;
 	uint16_t id;
-	uint16_t quanta;
+	uint64_t quanta;
 	uint8_t priviledge;
 	AuSigHandler singals[39];
 	SignalQueue* signalQueue;
