+++ Process/Deodhai/backdirty.h
@@ -57,9 +57,9 @@ extern void BackDirtyAdd(int x, int y, int w, int h);
 * @param h -- mem pointer where to store h
 * @param index -- index
 */
-extern void BackDirtyGetRect(int *x, int *y, int *w, int *h, uint8_t index);
+extern void BackDirtyGetRect(int *x, int *y, int *w, int *h, int index);
 
-extern uint8_t BackDirtyGetDirtyCount();
+extern int BackDirtyGetDirtyCount();
 
 extern void BackDirtyCountReset();
